moved the query frame into it's own module.
